requirements.txt
Django==3.1.5
pyjwt==2.0.1
djangorestframework-simplejwt==4.6.0
python-decouple==3.4
# Swagger
django-rest-swagger==2.2.0
channels==3.0.3
